refactor: separate ROIFilter component from ROITable
